split editor scroll handling
